+++ NEWS.md
@@ -1,3 +1,6 @@
 # resc (development version)
 
+* Add `get_countrycode()` to return the associated code with a country name
+* Add `get_countryname()` to return the associated name with a country code
+* Add `is_valid_country()` returns whether input as name or code is valid within Eurovision
 * Add `get_years()` which returns possible years pertaining to Eurovision